+++ src/modules/auth/number/PhoneNumberScreen.js
@@ -11,8 +11,12 @@ import {
 } from 'reactstrap';
 
 import { factorChaining } from "../utils/factorChaining";
+import generatePenrose from '../utils/generatePenrose';
 
 function PhoneNumberScreen(props) {
+
+	generatePenrose();
+	
 	return (
 		<Container>
 			<Row className="justify-content-center">